amx_exec.c, amx_verifier.c: Add basic compatibility for 'lctrl 7' and 'lctrl/sctrl 8'
lctrl 7 - PRI=1 if JIT is enabled, otherwise PRI isn't modified.
lctrl 8 - translates the code offset in PRI into a physical address and stores it in PRI if JIT is enabled and the cell size is big enough to fit a pointer (does nothing otherwise).
sctrl 8 - jumps to a physical address in PRI. This is equal to 'sctrl 6' if JIT isn't present.